requirements.txt
redis==4.4.2
elasticsearch[async]==7.9.1
fastapi==0.61.1
orjson==3.5.1
uvicorn==0.12.2
pre-commit==3.1.1

charset-normalizer==3.0.1
psycopg2==2.9.5
pydantic==1.10.5
python-dotenv==1.0.0
requests==2.28.2
parameterized==0.8.1

uvloop==0.17.0; sys_platform != "win32" and implementation_name == "cpython"